feat(phase 2): SQLite + sqlite-vec backend
Same surface as the postgres backend, dispatched by HEIRLOOM_BACKEND.
The Tauri build will set HEIRLOOM_BACKEND=sqlite at runtime.

- src/lib/db/sqlite.ts — better-sqlite3 + sqlite-vec adapter that
  mimics postgres.js's tagged-template API (callable sql, sql.begin,
  sql.unsafe, sql.json, transactional withRls)
- migrations/sqlite/001_schema.sql — full schema translated from the
  postgres handoff: TEXT primary keys with a DEFAULT (gen_uuid()),
  BLOB embeddings, JSON-as-TEXT, no RLS (single-user)
- sqlite-vec is loaded as an extension; vector columns use
  vec_distance_cosine() under the hood
- Two user functions registered on connect: now() and gen_uuid()
  (canonical UUIDv4 so the existing zod .uuid() validators still pass)
- Portability fixes across call sites: COUNT(*)::int → CAST(... AS
  INTEGER); LEFT(x, n) → substr(x, 1, n); ARRAY_AGG(... ORDER BY)[1] →
  MIN(...); NULL::numeric → NULL; renamed "exists" SQL alias since
  SQLite reserves it
- serverExternalPackages: ["better-sqlite3", "sqlite-vec"] so
  Turbopack doesn't try to bundle the native node addons

Both backends verified visually end-to-end: bootstrap → onboard →
note capture → reflection with citation.

+++ src/lib/letter-conditions.ts
@@ -210,7 +210,7 @@ async function fetchPending(
   }
   return tx<PendingRow[]>`
     SELECT id, capture_id, to_nominee_id, occasion_prompt, conditions,
-           NULL::numeric AS intent_similarity
+           NULL AS intent_similarity
       FROM sealed_letters
      WHERE vault_id = ${vault_id}
        AND unlocked_at IS NULL
